8-debris-federico-giuseppe-paolo-sandro/src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "numpy>=2.3.3",
    "peewee>=3.18.2",
    "pillow>=11.3.0",
    "pyserial>=3.5",
]
